V16
Some story added. There was a massive glitch that was caused by renaming the project in the file, not renaming it by saving it. If objects don't function according to the code and randomly turn into other objects it is not a problem with the objects in question, it is a problem with the file name and it is easy to fix.